Also set a stable time func for store in tests
The test suite is currently failing because although we'd accounted for
eventual board expiry at our stable time function in the test server, we
weren't pushing a stable time down into the store, so it was using the
real clock and immediately considering stored content invalid as it
had passed 22 days in age.

Here, also inject stable time into the test store to fix the problem.